Split integration test container setup into services module
Move the container lifecycle (start_*/create_* helpers, network and log
handling) from the integration conftest into tests/integration/services.py.
The conftest only wires pytest fixtures now, and the package conftests
import their query-runner helpers from the new module. No behavior change.
