+++ tests/integration/procedure_surface/arrow/node_embedding/conftest.py
@@ -5,16 +5,13 @@ import pytest
 from testcontainers.core.network import Network
 
 from graphdatascience.arrow_client.authenticated_flight_client import AuthenticatedArrowClient
-from tests.integration.conftest import (
+from tests.integration.services import (
+    RUNTIME_SESSION_ALIAS,
     GdsSessionConnectionInfo,
     create_arrow_client,
     start_runtime_api,
     start_session,
 )
-
-# Distinct from the default "gds-session" so this runtime-backed session can coexist with the
-# shared session on the same package network without a DNS alias collision.
-RUNTIME_SESSION_ALIAS = "gds-session-with-runtime"
 
 
 @pytest.fixture(scope="package")
@@ -36,13 +33,18 @@ def session_connection_runtime(
         tmp_path_factory,
         network,
         request,
-        runtime_api_uri=runtime_api,
         gds_api_uri=gds_api_connection,
+        runtime_api_uri=runtime_api,
         session_alias=RUNTIME_SESSION_ALIAS,
     )
 
 
 @pytest.fixture(scope="package")
 def arrow_client_runtime(session_connection_runtime: GdsSessionConnectionInfo) -> AuthenticatedArrowClient:
-    """Arrow client backed by a session wired to the python-runtime API (needed by FastPath)."""
+    """Arrow client backed by a session wired to the python-runtime API (needed by FastPath).
+
+    Package-scoped on purpose: the runtime-backed session and its mock runtime API are
+    stopped again once this package's tests are done, instead of idling until the end of
+    the whole test session.
+    """
     return create_arrow_client(session_connection_runtime)
