+++ tests/integration/procedure_surface/cypher/conftest.py
@@ -4,7 +4,7 @@ import pytest
 from testcontainers.neo4j import Neo4jContainer
 
 from graphdatascience.query_runner.neo4j_query_runner import Neo4jQueryRunner
-from tests.integration.conftest import create_plugin_query_runner
+from tests.integration.services import create_plugin_query_runner
 
 
 @pytest.fixture(scope="package")
