+++ tests/integration/procedure_surface/session/conftest.py
@@ -4,7 +4,7 @@ import pytest
 
 from graphdatascience.query_runner.neo4j_query_runner import Neo4jQueryRunner
 from graphdatascience.session.dbms_connection_info import DbmsConnectionInfo
-from tests.integration.conftest import create_db_query_runner
+from tests.integration.services import create_db_query_runner
 
 
 @pytest.fixture(scope="package")
